requirements.txt
PyYAML==6.0
gpiozero==1.6.2
flickrapi==2.4.0
Pillow==8.4.0
selenium==4.1.2
